Ignore implicit default persistence units if there is no datasource and no entity
... instead of failing as soon as there is a managed class, even if that
class is not an entity.

The previous behavior was particularly problematic when Panache was in the
classpath, because then there will always be a managed class (Panache's
mappedsuperclasses) in the default persistence unit, so we would end up
failing with default configuration.

Now we will just ignore such a persistence unit.