ArC: add validation of Instance/Event type argument
`Event` may have a wildcard type argument, but it must have a lower bound.
Similarly, `Instance` may have a wildcard type argument, but it must NOT have
a lower bound.

Further, this commit fixes the set of bean types of the `Instance` bean.
It doesn't follow the spec to the letter, because the `Instance` and `Provider`
bean types should be parametererized, but it's good enough to pass the TCK.