Fix bootstrap when module is pointing to POM file directly
Also make sure the Map cache in WorkspaceLoader is consistently
referencing the POM files.

Note that I stumbled upon another limitation: the app artifact requires
a strict pom.xml as assumptions are hardcoded in
LocalProject#locateCurrentProjectPom().
I haven't fixed that one as it probably requires a lot more changes.

I added a test to make sure we don't regress on this.

Fixes #50571