Maven ITs - Reduce waiting times in multi-module tests
We were having a very long waiting time in multi-module tests due to the
dependency module not having a class.
This was throwing an exception and then we were waiting forever before
actually doing something else.

The fact that the project is empty is somehow related to the test I
removed. The test was introduced in
https://github.com/quarkusio/quarkus/pull/9197 to check we didn't have a
warning when there were no Java sources. Given we actually throw an
exception in the multi-module project now if the classes are not around,
it doesn't make a lot of sense to check for this warning anyway.

(cherry picked from commit 7dc16c015e8267ae78ee1df78d57d459983b7a79)