Pass app model location to Maven test process
This aligns the behavior with the Gradle plugin and speeds up startup
because we no longer have to search for a POM file to find the workspace
location and load the app model.

It also makes Quarkus out-of-the-box compatible with Develocity test
distribution, because POMs are generally not transferred as inputs to the
distributed test workers.