Expose the test deploy classpath to Develocity
In order to run tests on remote agents, Develocity needs to know all files
necessary to run the tests, as well as the application under test. The deploy
classpath contains dependencies that are not part of the regular Maven project
model and we need to expose it in some way that Develocity can easily grab and
track it.