Fix version resolution in  CLI
This was broken in #49926 where the property name was changed
mistakenly.

To be honest, I'm also not sure the new method is as safe as it used to
be, especially on Windows as we also dropped the specific handling of
the jar pattern. There was a specific comment about it in the
preexisting method.
I suppose we will see how it goes...