Fix Amazon Lambda handler method resolution
The handler method resolution was relying on the order in which methods
were returned by getMethods(), which is not guaranted at all.
This makes the extension fail with IBM Semeru for LambdaWithHierarchyTest.

The new implementation should be a lot more solid but is also doing a
lot more reflection calls.
My guess is that we should probably move this resolution to Jandex