Proxy Registry: big refactoring
The public API is in the `io.quarkus.proxy` package and consists of:

- `ProxyType`: an enum of proxy types (HTTP, SOCKS4, SOCKS5)
- `ProxyConfiguration`: an interface that provides access to configured proxy
- `ProxyConfigurationRegistry`: an interface that allows looking up a named
  proxy configuration, optimized for the most common usage

The `io.quarkus.proxy.runtime` package contains implementation, which is:

- `ProxyConfigurationRecorder`: Quarkus recorder that provides a `Supplier`
  that is used in 2 ways: a proxy registry carrier stored in a build item,
  and a recorder proxy for a `ProxyConfigurationRegistry` synthetic bean
- `ProxyConfigurationRegistryImpl`: implementation of `ProxyConfigurationRegistry`
  that is created by the recorder

The config interface is now in the `io.quarkus.proxy.runtime.config`, because
the interface itself is not a public API.

The `deployment` module gains a `ProxyRegistryBuildItem` that can be consumed
by build steps that need to invoke recorder methods with proxy configuration.
The `ProxyRegistryProcessor` hasn't changed much (except of the name).

Finally, this commit adds a few basic unit tests to the Proxy Registry extension.