Use simplified AOT file creation workflow provided by JEP-514
Quarkus had support for generating the JDK's AOT file
using the two-step workflow specified in JEP-483.
With the advent of JDK 25 however, there is now
a simpler workflow for creating the AOT file which
has been specified by JEP-514.
This change shouldn't have any effect on users (except for requiring JDK 25 for the AOT file),
but it does simplify the Quarkus code