ArC: fix searching for decorated methods
In case multiple decorated types of a single decorator declare the same
method (it may have the same signature, but in presence of generic types,
it doesn't have to), we used to remember all of them, which later lead
to an error when generating the decorating subclass.

This commit fixes that: we really only need to remember _one_ method
from each decorator, and it doesn't really matter which exact one it is,
because all decorated methods are virtual and we invoke them as such.