Download all dependencies by default when going offline
It is sometimes confusion and cumbersome to explicitly call the right mode before going offline. For example if you are about to board a plane you might forget to give the option to also download dependencies for dev mode. (I actually had that last week).

This introduces a new option that downloads dev, test and prod dependencies and makes this the new default. This allows developers to go offline without having to specify the mode explicitly.

Furthermore, adjust the usage through the docs and CI to align with the previous state. For building native images we don't need the dev and test dependencies. For running on CI we don't need the dev dependencies either.

(cherry picked from commit 7f514f5d594379fa184151d758d7778fda3c1179)