ArC: fix injection into non-app beans with app decorators
A while ago, we turned generated `_Bean`, `_Subclass` and `_ClientProxy`
classes pertaining to non-app beans into app classes to allow decoration
by app decorators. This in turn made injection into non-`public` IPs
impossible.

We fix it in this commit by a bytecode transformation that makes these IPs
`public`. However, this might break subclasses (albeit in relatively rare
cases), so in case the bean class has subclasses, we error out.

Also, this commit turns the generated `_ClientProxy` classes back into
non-app, because that allows invoking non-`public` methods. Turning
the generated `_ClientProxy` class into app class was arguably a mistake.

Finally, this commit removes the `<scope>test</scope>` on 2 dependencies
in the `extensions/arc/deployment` module, because when test-scoped
dependencies are used in `QuarkusUnitTest.setForcedDependencies()`,
they are treated as application archives. If they have the default scope,
they are treated as non-app archives, which is exactly what we need
in all tests in this module that use `setForcedDependencies()`. To make
sure we don't regress again, a comment is added in the POM and all tests
that use this feature also assert that the classes from the external
archives are indeed loaded by the base runtime CL (and not runtime CL).