Run quarkus-hibernate-orm-rest-data-panache DevMode tests in a separate Surefire execution
These tests have been flaky due to OOM: Metaspace lately, see
https://ge.quarkus.io/scans/tests?search.timeZoneId=Europe%2FParis&tests.container=io.quarkus.hibernate.orm.rest.data.panache.deployment.build.BuildConditionsWithResourceDisabledTest&tests.test=#

I believe we have the same problem here as in the Hibernate ORM
extension: many devmode tests, combined with a metaspace leak when
using QuarkusDevModeTest, leaking to OOM.

It's possible more investigation is needed, in particular because this
OOM is new.
If anyone thinks so, maybe we should create an issue and work on that
whenever possible.

In the meantime... this will at least avoid flaky CI.