Load Shedding: use RoutingContext instead of HttpServerRequest
This is because `RoutingContext`, unlike `HttpServerRequest`, has a method
`normalizedPath()`, which we need to use in `ManagementRequestPrioritizer`.
Previously, we used `HttpServerRequest.path()`, which returns the path
unmodified, so the request prioritizer considers `/q/../something` as
pertaining to the non-application endpoint, even though it is not.
`RoutingContext.normalizedPath()` returns `/something` in this case,
as expected.

Further, this commit somewhat improves the load shedding documentation.