ArC: guard against multiple occurrences of the same decorator
Gizmo 1 didn't bother checking that multiple fields of the same name are added
to a class, so `SubclassGenerator` didn't fail in case the same decorator
occurred multiple times in the `BeanInfo.getNextDecorators()` map. However,
Gizmo 2 does check, so we have to guard against it. The guard is simple:
instead of adding decorators to a list directly, we first add them to a set
and when that's done, we create a list from that set.

(cherry picked from commit ac46b3133f9e73ed3ffc23e88e2cbcaed2aa022f)