Set expiresIn to one hour for OIDC Dev Services
To match with the `expires_in` that we manually set for the token responses. This should make debugging the property `quarkus.oidc.authentication.session-age-extension` a bit easier, as you now get a token that's valid for a day.

We can probably also make this configurable, as that would be useful if you want to test locally what happens if your token expires. Will add that in a next commit.

(cherry picked from commit 6b9211ca3c4c820c80636d8758895f26988941d9)