Keep track of generated and transformed classes to avoid lookups
It happens quite often that a given directory has a couple of
generated/transformed classes and in this case we hit the
generated/transformed-bytecode jars for each lookup in this directory,
even for the regular case of the class being in the "standard" jars.

Keeping track of generated and transformed classes allows to avoid this
additional work with minimal added cost.

In my testing, it dropped 10 ms from the startup time, going from 367 to
353ms for a simple Quarkus REST app.