Compute bean raw types in a Set/HashSet before materializing them
It allows to avoid performing very badly when we have lots of beans of
the same type, for instance when you have a lot of Panache repositories
for various entities.

We materialize things with Lists in the end, and they might be a bit more
optimized than before for small non-singleton lists.

The downside is that we have a bit more allocations but from my testing,
it wasn't affecting startup times.

Fixes #51468