ArC: materialize the map bean ID -> bean eagerly
When loading beans from `ComponentsProvider`s, we iterate through all of them
in order to store them in the right lists. When doing that, we have a chance
to load the ID (which is always a constant string) and eagerly build the map
of bean ID -> bean. Everything stored in the map is on the heap already, so
the only overhead is the map structure itself. This is fairly small, especially
since the map is later copied into an immutable map using `Map.copyOf()`, and
the immutable map is just a simple array with linear probing.