Do not propagate default db-kind in Dev Services properties
It can lead to this sort of errors:
quarkus.datasource.db-kind is set to 'postgresql' but it is build time fixed to 'null'. Did you change the property quarkus.datasource.db-kind after building the application?
when the db-kind is not set and we are using the default one,
because it ends up being passed as a system property to the app we start
in the ITs.