Require Maven 3.9.0+ for Quarkus projects
Maven 3.9.0 has been released on Jan 31st 2023 so it will be two years
when we will release 3.31.
I think we can now mandate it.
We are using more and more advanced features of Maven and I think we
need to set a reasonable minimum version given we don't test 3.8 at all.